requirements.txt
flask==3.0.0
vosk==0.3.45
pyaudio==0.2.14
requests==2.31.0
eventlet==0.33.3
flask-socketio==5.3.6
python-socketio==5.10.0
werkzeug==3.0.1
numpy==1.26.2
gsv-tts-lite==0.3.5
